Add tests for round-tripping ByteStrings through the CBString API.
It's of vital importance that the CBString API be capable of round-tripping
strings. If it isn't, then data stored in HyperDex won't be equivalent after
being retrieved. Moreso, the HyperDex API relies on binary serialization of
data types and a failure to serialize the data correctly may break the API.