Fixed roundtripping integers through HyperDex
This was a silly mistake - the client was opened using "withClient" but
closed prematurely. The result was that after a short time the test
would fail, a sort of Heisenbug.

As well, a bug (?) was fixed in the Client.chs loopClient function that
caused looping for any nonpresent handle not to terminate. Now, any
handle not in the map is considered already complete.