Demonstrate successful searches when using a warm client.
In this test, as opposed to the previous commit, the hyperdex_search
call is run using a warm client from a pool of clients previously used
for other tests. The testSearch test is the last one executed, and the
success of these tests can be demonstrated by running:

for I in {1..1000}; do ./dist/build/tests/tests -a 1 -t atomic >> ./testresults; done

Searching the output for HyperclientInterrupted or HyperclientPollfailed - the
Haskell names for each return code previously reported - yields no results.